refactor: Reservation의 Id Wrapper class 추가
