+++ src/main/java/roomescape/reservation/domain/ReservationId.java
@@ -1,0 +1,5 @@
+package roomescape.reservation.domain;
+
+public record ReservationId(long id) {
+
+}
